feat: ask more than one question
Instead of reading the arguments passed and use them as a "prompt" for GPT,
we prompt the user and scan their input. We also keep the previously
asked questions and send them each time a request is made to make it
seem like a conversation. Will need to handle API rate limitting at some
point and make it testable.